refactor(Client): move `Client` into its own file
